refactor(models, dao): Relocate `Branch` and `Property` models for better modularity
- Moved `Branch` and `Property` to `models` package to align with package structure.
- Deleted outdated `BranchDao` and its dependencies.
- Updated references in `SuperAdminController` to reflect new paths.
